+++ docs/docs/news.md
@@ -16,6 +16,7 @@ A new family of `update_*` functions to easily update graph aspects on the fly.
 - `update_node_size` and `update_node_color` to update a single node.
 - `update_nodes_size` and `update_nodes_color` to update multiple nodes from a data.frame.
 - `update_link_source_color` and `update_link_target_color` to update a link source and target color.
+- `update_links_source_color` and `update_links_target_color` to update multiple links from a data.frame.
 
 # 0.1.0